# api-gateway/internal/http/docs/swagger.yaml
definitions:
  genproto.AttachmentCreateRes:
    properties:
      file_url:
        type: string
    type: object
  genproto.AttachmentGetAllRes:
    properties:
      attachments:
        items:
          $ref: '#/definitions/genproto.AttachmentGetRes'
        type: array
      pagination:
        $ref: '#/definitions/genproto.Pagination'
    type: object
  genproto.AttachmentGetRes:
    properties:
      file_name:
        type: string
      file_size:
        type: string
      file_url:
        type: string
      id:
        type: string
      mime_type:
        type: string
      outbox:
        $ref: '#/definitions/genproto.OutboxMessageGetRes'
    type: object
...
  genproto.ConfirmUserReq:
    properties:
      code:
        type: string
      email:
        type: string
    type: object
...
  genproto.LoginReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  genproto.MessageSendBCC:
    properties:
      email:
        items:
          type: string
        type: array
    type: object
  genproto.MessageSendCC:
    properties:
      email:
        items:
          type: string
        type: array
    type: object
  genproto.MessageSendTo:
    properties:
      email:
        items:
          type: string
        type: array
    type: object
...
  genproto.OutboxMessageGetRes:
    properties:
      attachment_ids:
        items:
          type: string
        type: array
      body:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      is_archived:
        type: boolean
      is_draft:
        type: boolean
      is_starred:
        type: boolean
      receivers:
        $ref: '#/definitions/genproto.Receivers'
      sender:
        $ref: '#/definitions/genproto.UserGetRes'
      sent_at:
        type: string
      subject:
        type: string
    type: object
...
  genproto.Pagination:
    properties:
      limit:
        type: integer
      skip:
        type: integer
    type: object
  genproto.Receivers:
    properties:
      bcc:
        $ref: '#/definitions/genproto.MessageSendBCC'
      cc:
        $ref: '#/definitions/genproto.MessageSendCC'
      to:
        $ref: '#/definitions/genproto.MessageSendTo'
    type: object
...
  genproto.UserCreateReqForSwagger:
    properties:
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
  genproto.UserGetRes:
    properties:
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      pfp_url:
        type: string
    type: object
...
  token.Tokens:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
info:
  contact: {}
paths:
  /attachment:
    post:
      consumes:
      - multipart/form-data
      description: Creates a new attachment associated with an outbox message.
      parameters:
      - description: Outbox message ID
        in: formData
        name: outbox_id
        required: true
        type: string
      - description: Attachment file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Attachment created successfully
          schema:
            $ref: '#/definitions/genproto.AttachmentCreateRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new attachment
      tags:
      - 07-Attachments
  /attachment/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an attachment by its ID.
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Attachment deleted successfully
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Attachment not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete an attachment
      tags:
      - 07-Attachments
  /attachment/{outbox_id}:
    get:
      consumes:
      - application/json
      description: Retrieves all attachments associated with an outbox message.
      parameters:
      - description: Outbox message ID
        in: path
        name: outbox_id
        required: true
        type: string
      - description: Filter by MIME type
        in: query
        name: mime_type
        type: string
      - description: Filter by creation date (from)
        in: query
        name: created_from
        type: string
      - description: Filter by creation date (to)
        in: query
        name: created_to
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of attachments per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Attachments retrieved successfully
          schema:
            $ref: '#/definitions/genproto.AttachmentGetAllRes'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Attachments not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get attachments for an outbox message
      tags:
      - 07-Attachments
  /confirm-registration:
    post:
      consumes:
      - application/json
      description: Confirms a user's registration using the code sent to their email.
      parameters:
      - description: Confirmation request
        in: body
        name: confirmation
        required: true
        schema:
          $ref: '#/definitions/genproto.ConfirmUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: JWT tokens
          schema:
            $ref: '#/definitions/token.Tokens'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Incorrect verification code
          schema:
            type: string
        "404":
          description: Verification code expired or email not found
          schema:
            type: string
      summary: Confirm registration with code
      tags:
      - 01-Authentication
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/genproto.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: JWT tokens
          schema:
            $ref: '#/definitions/token.Tokens'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Invalid email or password
          schema:
            type: string
      summary: Login a user
      tags:
      - 01-Authentication
  /logout:
    post:
      consumes:
      - application/json
      description: Logs out the user by blacklisting their refresh token.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Logout a user
      tags:
      - 01-Authentication
...
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with email, username, and password
      parameters:
      - description: User registration request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/genproto.UserCreateReqForSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: JWT tokens
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Register a new user
      tags:
      - 01-Authentication
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...